auth/requirements.txt
astroid==2.12.13
black==22.10.0
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-API==3.0.post1
Flask-MySQLdb==1.0.1
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
parso==0.8.3
pathspec==0.10.2
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.6
python-dotenv==0.21.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
